Remove package file
