+++ modules/ingestor/src/main/scala/mongo.chapter.scala
@@ -11,6 +11,8 @@ import mongo4cats.database.MongoDatabase
 import mongo4cats.operations.{ Accumulator, Aggregate, Filter }
 import org.typelevel.log4cats.Logger
 
+import Repo.*
+
 trait ChapterRepo:
   // Aggregate chapters data and convert them to StudyChapterText by their study ids
   def byStudyIds(ids: List[String]): IO[Map[String, StudyData]]
