+++ modules/ingestor/src/main/scala/mongo.game.scala
@@ -20,7 +20,7 @@ import org.typelevel.log4cats.{ Logger, LoggerFactory }
 import java.time.Instant
 import scala.concurrent.duration.*
 
-import Repo.Result
+import Repo.{ *, given }
 
 object GameRepo:
