+++ modules/ingestor/src/main/scala/mongo.team.scala
@@ -15,7 +15,7 @@ import org.typelevel.log4cats.{ Logger, LoggerFactory }
 import java.time.Instant
 import scala.concurrent.duration.*
 
-import Repo.Result
+import Repo.{ *, given }
 
 object TeamRepo:
